refactor(web): dedup shared filter api types
Extract the duplicated Device, VlanRange, IPNet and ListConfigsResponse
wire types into a single api/shared.ts and re-export them from the
forward, acl, fwstate and routes api modules.
